+++ R/bind_remstats.R
@@ -104,6 +104,8 @@ bind_remstats <- function(...) {
         attr(statistics, "model") <- attr(arg.list[[1]], "model")
         attr(statistics, "formula") <- final_formula
         attr(statistics, "riskset") <- attr(arg.list[[1]], "riskset")
+        attr(statistics, "subset") <- attr(arg.list[[1]], "subset")
+        attr(statistics, "method") <- attr(arg.list[[1]], "method")
         attr(statistics, "adjmat") <- attr(arg.list[[1]], "adjmat")
 
         # Output
@@ -255,6 +257,8 @@ bind_remstats <- function(...) {
         class(statistics) <- class(arg.list[[1]])
         attr(statistics, "model") <- attr(arg.list[[1]], "model")
         attr(statistics, "actors") <- attr(arg.list[[1]], "actors")
+        attr(statistics, "subset") <- attr(arg.list[[1]], "subset")
+        attr(statistics, "method") <- attr(arg.list[[1]], "method")
 
         # Output
         return(statistics)