+++ src/RcppExports.cpp
@@ -15,14 +15,14 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // getEventIndices
-Rcpp::List getEventIndices(const arma::mat& edgelist, int start, int stop, std::string method, std::string model);
+Rcpp::List getEventIndices(const arma::mat& edgelist, int start, int stop, Rcpp::String method, std::string model);
 static SEXP _remstats_getEventIndices_try(SEXP edgelistSEXP, SEXP startSEXP, SEXP stopSEXP, SEXP methodSEXP, SEXP modelSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::traits::input_parameter< const arma::mat& >::type edgelist(edgelistSEXP);
     Rcpp::traits::input_parameter< int >::type start(startSEXP);
     Rcpp::traits::input_parameter< int >::type stop(stopSEXP);
-    Rcpp::traits::input_parameter< std::string >::type method(methodSEXP);
+    Rcpp::traits::input_parameter< Rcpp::String >::type method(methodSEXP);
     Rcpp::traits::input_parameter< std::string >::type model(modelSEXP);
     rcpp_result_gen = Rcpp::wrap(getEventIndices(edgelist, start, stop, method, model));
     return rcpp_result_gen;
@@ -53,7 +53,7 @@ RcppExport SEXP _remstats_getEventIndices(SEXP edgelistSEXP, SEXP startSEXP, SEX
     return rcpp_result_gen;
 }
 // compute_stats_sender
-arma::cube compute_stats_sender(Rcpp::CharacterVector& effects, const arma::mat& edgelist, const arma::vec& actors, const arma::vec& weights, const Rcpp::List& covariates, const Rcpp::List& interactions, std::string memory, const arma::vec memory_value, Rcpp::CharacterVector& scaling, int start, int stop, std::string method, bool display_progress);
+arma::cube compute_stats_sender(Rcpp::CharacterVector& effects, const arma::mat& edgelist, const arma::vec& actors, const arma::vec& weights, const Rcpp::List& covariates, const Rcpp::List& interactions, Rcpp::String memory, const arma::vec memory_value, Rcpp::CharacterVector& scaling, int start, int stop, Rcpp::String method, bool display_progress);
 static SEXP _remstats_compute_stats_sender_try(SEXP effectsSEXP, SEXP edgelistSEXP, SEXP actorsSEXP, SEXP weightsSEXP, SEXP covariatesSEXP, SEXP interactionsSEXP, SEXP memorySEXP, SEXP memory_valueSEXP, SEXP scalingSEXP, SEXP startSEXP, SEXP stopSEXP, SEXP methodSEXP, SEXP display_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -63,12 +63,12 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::vec& >::type weights(weightsSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List& >::type covariates(covariatesSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List& >::type interactions(interactionsSEXP);
-    Rcpp::traits::input_parameter< std::string >::type memory(memorySEXP);
+    Rcpp::traits::input_parameter< Rcpp::String >::type memory(memorySEXP);
     Rcpp::traits::input_parameter< const arma::vec >::type memory_value(memory_valueSEXP);
     Rcpp::traits::input_parameter< Rcpp::CharacterVector& >::type scaling(scalingSEXP);
     Rcpp::traits::input_parameter< int >::type start(startSEXP);
     Rcpp::traits::input_parameter< int >::type stop(stopSEXP);
-    Rcpp::traits::input_parameter< std::string >::type method(methodSEXP);
+    Rcpp::traits::input_parameter< Rcpp::String >::type method(methodSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(compute_stats_sender(effects, edgelist, actors, weights, covariates, interactions, memory, memory_value, scaling, start, stop, method, display_progress));
     return rcpp_result_gen;
@@ -99,7 +99,7 @@ RcppExport SEXP _remstats_compute_stats_sender(SEXP effectsSEXP, SEXP edgelistSE
     return rcpp_result_gen;
 }
 // compute_stats_receiver
-arma::cube compute_stats_receiver(Rcpp::CharacterVector& effects, const arma::mat& edgelist, const arma::vec& actors, const arma::vec& weights, const Rcpp::List& covariates, const Rcpp::List& interactions, std::string memory, const arma::vec memory_value, Rcpp::CharacterVector& scaling, int start, int stop, std::string method, bool display_progress);
+arma::cube compute_stats_receiver(Rcpp::CharacterVector& effects, const arma::mat& edgelist, const arma::vec& actors, const arma::vec& weights, const Rcpp::List& covariates, const Rcpp::List& interactions, Rcpp::String memory, const arma::vec memory_value, Rcpp::CharacterVector& scaling, int start, int stop, Rcpp::String method, bool display_progress);
 static SEXP _remstats_compute_stats_receiver_try(SEXP effectsSEXP, SEXP edgelistSEXP, SEXP actorsSEXP, SEXP weightsSEXP, SEXP covariatesSEXP, SEXP interactionsSEXP, SEXP memorySEXP, SEXP memory_valueSEXP, SEXP scalingSEXP, SEXP startSEXP, SEXP stopSEXP, SEXP methodSEXP, SEXP display_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -109,12 +109,12 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::vec& >::type weights(weightsSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List& >::type covariates(covariatesSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List& >::type interactions(interactionsSEXP);
-    Rcpp::traits::input_parameter< std::string >::type memory(memorySEXP);
+    Rcpp::traits::input_parameter< Rcpp::String >::type memory(memorySEXP);
     Rcpp::traits::input_parameter< const arma::vec >::type memory_value(memory_valueSEXP);
     Rcpp::traits::input_parameter< Rcpp::CharacterVector& >::type scaling(scalingSEXP);
     Rcpp::traits::input_parameter< int >::type start(startSEXP);
     Rcpp::traits::input_parameter< int >::type stop(stopSEXP);
-    Rcpp::traits::input_parameter< std::string >::type method(methodSEXP);
+    Rcpp::traits::input_parameter< Rcpp::String >::type method(methodSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(compute_stats_receiver(effects, edgelist, actors, weights, covariates, interactions, memory, memory_value, scaling, start, stop, method, display_progress));
     return rcpp_result_gen;
@@ -324,9 +324,9 @@ END_RCPP
 static int _remstats_RcppExport_validate(const char* sig) { 
     static std::set<std::string> signatures;
     if (signatures.empty()) {
-        signatures.insert("Rcpp::List(*getEventIndices)(const arma::mat&,int,int,std::string,std::string)");
-        signatures.insert("arma::cube(*compute_stats_sender)(Rcpp::CharacterVector&,const arma::mat&,const arma::vec&,const arma::vec&,const Rcpp::List&,const Rcpp::List&,std::string,const arma::vec,Rcpp::CharacterVector&,int,int,std::string,bool)");
-        signatures.insert("arma::cube(*compute_stats_receiver)(Rcpp::CharacterVector&,const arma::mat&,const arma::vec&,const arma::vec&,const Rcpp::List&,const Rcpp::List&,std::string,const arma::vec,Rcpp::CharacterVector&,int,int,std::string,bool)");
+        signatures.insert("Rcpp::List(*getEventIndices)(const arma::mat&,int,int,Rcpp::String,std::string)");
+        signatures.insert("arma::cube(*compute_stats_sender)(Rcpp::CharacterVector&,const arma::mat&,const arma::vec&,const arma::vec&,const Rcpp::List&,const Rcpp::List&,Rcpp::String,const arma::vec,Rcpp::CharacterVector&,int,int,Rcpp::String,bool)");
+        signatures.insert("arma::cube(*compute_stats_receiver)(Rcpp::CharacterVector&,const arma::mat&,const arma::vec&,const arma::vec&,const Rcpp::List&,const Rcpp::List&,Rcpp::String,const arma::vec,Rcpp::CharacterVector&,int,int,Rcpp::String,bool)");
         signatures.insert("arma::mat(*get_riskset)(arma::uvec,arma::uvec,bool)");
         signatures.insert("arma::mat(*convert_to_risksetMatrix)(arma::mat,int,int)");
         signatures.insert("arma::mat(*calculate_inertia)(const arma::mat&,const arma::vec&,const arma::mat&,Rcpp::String,const arma::vec&,int,int,bool,Rcpp::String)");